requirements.txt
amqp==2.5.2
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
boto3==1.19.1
botocore==1.22.1
Brlapi==0.7.0
cached-property==1.5.2
celery==4.4.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
clinner==1.12.3
colorama==0.4.3
colorlog==3.2.0
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
cryptography==35.0.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==2.1.3
Django==3.2.8
django-celery-beat==2.2.1
django-choices==1.7.2
django-cors-headers==3.10.0
django-filter==21.1
django-health-check==3.16.5
django-import-export==2.8.0
django-model-utils==4.2.0
django-rest-swagger==2.2.0
django-ses==2.3.0
django-storages==1.11.1
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djoser==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.8.0
Fiona==1.8.20
future==0.18.2
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.26
gunicorn==20.1.0
health-check==3.4.1
html5lib==1.1
httplib2==0.14.0
idna==2.8
imageio==2.16.0
importlib-metadata==4.10.0
iniconfig==1.1.1
install==1.3.5
ipython==7.28.0
isodate==0.6.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
monotonic==1.5
more-itertools==8.10.0
moto==2.2.10
munch==2.5.0
netifaces==0.10.4
networkx==2.6.3
numpy==1.22.1
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.0
pandas==1.4.0
paramiko==2.6.0
parso==0.8.2
pdfkit==0.6.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2022.1.8
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.21
protobuf==3.6.1
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyproj==3.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.14.0
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.19.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentry-sdk==1.5.4
Shapely==1.8.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
systemd-python==234
tablib==3.2.1
texttable==1.6.2
tifffile==2022.2.9
toml==0.10.2
traitlets==5.1.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.16.3
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.0.2
wkhtmltopdf==0.2
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
zeep==4.1.0
zipp==1.0.0
